Refactor TF common functions to utilities
