+++ scripts/pretrain_observer.py
@@ -9,36 +9,7 @@ from tensorflow.keras.callbacks import ModelCheckpoint
 
 from src.observers.spectrogram_observer import build_spectrogram_observer
 from src.utils.config_manager import Config
-
-
-class CustomWandbCallback(tf.keras.callbacks.Callback):
-    def on_epoch_end(self, epoch, logs=None):
-        wandb.log({"epoch": epoch + 1, "training_loss": logs.get("loss"), "validation_loss": logs.get("val_loss")})
-
-
-def parse_tfrecord(serialized_example):
-    """
-    Parse a single TFRecord into (spectrogram, param_error) Tensors.
-    """
-    feature_spec = {
-        'spectrogram': tf.io.VarLenFeature(tf.float32),
-        'param_error': tf.io.VarLenFeature(tf.float32),
-        'spectrogram_shape': tf.io.VarLenFeature(tf.int64),
-        'param_error_shape': tf.io.VarLenFeature(tf.int64),
-    }
-    parsed = tf.io.parse_single_example(serialized_example, feature_spec)
-
-    # Convert sparse to dense
-    spectrogram_flat = tf.sparse.to_dense(parsed['spectrogram'])
-    param_error_flat = tf.sparse.to_dense(parsed['param_error'])
-    spectrogram_shape = tf.sparse.to_dense(parsed['spectrogram_shape'])
-    param_error_shape = tf.sparse.to_dense(parsed['param_error_shape'])
-
-    # Reshape tensors to their original shapes
-    spectrogram = tf.reshape(spectrogram_flat, spectrogram_shape)
-    param_error = tf.reshape(param_error_flat, param_error_shape)
-
-    return spectrogram, param_error
+from src.utils.tensorflow import parse_tfrecord, CustomWandbCallback
 
 
 def main():
@@ -47,7 +18,7 @@ def main():
     # -------------------------------------------------------------------
     script_dir = os.path.dirname(os.path.abspath(__file__))
     config = Config()
-    config_path = os.path.join(script_dir, "configs", "train_observer.yaml")
+    config_path = os.path.join(script_dir, "configs", "pretrain_observer.yaml")
     config.load(config_path)
 
     # Store basic settings from config
